requirements.txt
asgiref==3.7.2
cloudinary==1.34.0
Django==3.2.21
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
urllib3==1.26.16
